Modify memory_effects rules and reported information for new refcount opts
Summary: Combine MayLoadStore and MayLoadStoreKill into a GeneralEffects tag,
and use helpers to create it in different situations.  Also adds a new field
that is required for refcount opts, which is locations that are loaded to be
"moved" to another location (without an incref, but taking the ownership of the
reference)---we have a few IR instructions that do this.  This class has no
import for load or store elim.  This diff also changes the rules for
GeneralEffects and PureLoad/PureStore to imply that if an instruction stores to
a location in GeneralEffects::stores, it does so with "normal memory
barriers"---i.e. it decrefs the old value that was there (if there was
one)---this means StElem must be PureStore for correctness (and LdElem can be
but it's ok as it was too).  LdLocPseudoMain still must /not/ be a PureLoad
because it's also a guard.  GenericRetDecRefs is weird and I tried to sorta
shoe-horn it into that ...

Reviewed By: @swtaarrs

Differential Revision: D1853640