Cleanup
Summary: Cleanup and refactor of the FastCGI application server. In D1816135 it became
apparant that the server has become frighteningly unmaintainable; no one seems
to understand the ownership or the concurrency model.

The Connection, Session, and Transaction classes have been merged into a single
FastCGISession class. This was done because the numerous callbacks between these
classes made it difficult to determine when the connection was shutting down,
and masked a number of latent use-after-free bugs.

Comments in fastcgi-session.h and fastcgi-transport.h explain the finer details
of ownership and concurrency. Most of the more subtle bugs were in the session
and connection logic, a new model has been introduced to better track pending
callbacks to avoid prematurely destroying the session object. Event counting
replaces write counts so that the pending EOF, and onComplete callbacks can be
tracked in addition to pending calls to writeSuccess and readErr. All locations
where the session is freed are done through destroy() calls as delete does not
respect the guards.

Most of the changes in fastcgi-transport.h were aesthetic, although a number of
places called lock and unlock directly on a thrift monitor object. The locking
has been rewritten to use RAII lock guards, and the POST data fetching has been
simplified.

The new FastCGI implementation supports the FastCGI flag FASTCGI_KEEP_CONN
to keep connections between the webserver and application server open. The
implementation behaves the same way as php-fpm and has been tested with nginx.
Unfortunately, a performance gain was not seen by enabling this flag in the
oss-performance utility. The flag also seemed to have a negative effect on
availability though this may have been due to an nginx bug with connection
persistence.

Benchmarking showed no major effect on performance, though for most frameworks
the number of non-200 reponses was substantially lower, demonstrating an
increased availability.

{sync, type="parent", child="external", childrevid="33111", childdiffid="173145", parentdiffid="6477847"}

Reviewed By: @jwatzman

Differential Revision: D1833845