Use std::vector instead of TinyVector in util/arena.h
Summary: The optimization to inline one of the pointers only helps with how
quickly arenas can be freed, since we hoist the `current' pointer.  We only use
these Arenas in the JIT and Verifier now (they used to be used for some
formerly hot runtime structures like VarEnvs), so it seems like we don't need
to worry about it these days (or worry about sizeof(Arena) for that matter).
The risk with the TinyVector is also that it limits the number of slabs to
2^16-1, and handling of the overflow condition is just a dbg-mode only assert
in CompactSizedPtr.

Reviewed By: @edwinsmith

Differential Revision: D1822164