Add hh_client extension and \HH\Client\typecheck()
Summary: The bifurcation between HHVM and `hh_client` can be super
confusing for new Hack engineers. This is the beginning of a plan to
unify them more closely. This diff provides two new functions that will
call `hh_client` for you and report its result, as if it were all of
HHVM itself.

The ultimate plan is that there will be two use cases.

- The basic use case. We'll automatically call `\HH\Client\typecheck_and_error()` in non-repo-auth requests the first time they `require` a Hack file (or something like that).
- The FB ("advanced") use case. The above will be turned off via INI option. Instead, after our init stack runs, we'll call `\HH\Client\typecheck()` and deal with the result ourselves (probably by manually calling `->triggerError(E_RECOVERABLE_ERROR, E_WARNING)`). This allows our nice "big red box" error handling to take effect, same for other OSS projects that are similarly advanced.

Or something like that. I'm sure the details will continue to be fleshed
out and change as we go. But that's where I'm ultimately going with
this.

NB: the API here is far from stable and will almost certainly change as
we go. It's undocumented and experimental -- @​simon_w, I'm looking at
you.

{sync, type="parent", child="external", childrevid="35859", childdiffid="187263", parentdiffid="6950451"}

Reviewed By: @paulbiss

Differential Revision: D1940190