Fix a bug in InterpOne of FPassM
Summary: FPassM might or might not affect the types of locals, depending on
whether the target function takes its arg by reference.  If it wasn't by
reference, and the FPassM had a local base and a prop dim, interp one would
assume it stdClass-promoted even though it didn't, and we'd crash when the
local is accessed later.

Reviewed By: @alexmalyshev

Differential Revision: D1822589