Allow Typeconsts to be Declared in Interfaces
Summary: Now that most of the implementation detail for type constants are
settled I able allowing them to be defined in interfaces. We delayed
adding support initially since it was not clear how we would resolve
conflicting definitions of type constants. This problem is not unique to
type constants so I do what we do for other abstract members.

I improved the error message for conflicting definitions in general by
detecting when these conflicts occur and cautioning the user to
redeclare the member.

Reviewed By: @dabek

Differential Revision: D1825740