Remove "log" from dfind env
Summary: This doesn't make any sense, to have a separate log; just use
stdout, which is already redirected to the Hack log file. This removes a
call to `Sys_utils.logname`, which I'm trying to kill, so that
communication and operation of the server is independent of the current
user.

{sync, type="parent", child="external", childrevid="32775", childdiffid="169875", parentdiffid="6374845"}

Reviewed By: @int3, @elgenie

Differential Revision: D1823509