Delete ClientStatus
Summary: This doesn't work anyways. I could fix it, but the fact that
it's been broken for a long time indicates that it isn't worth the
effort, especially since I want to rework this code to not have a notion
of "user" and this code has that assumption baked into it deeply.

{sync, type="parent", child="external", childrevid="32859", childdiffid="170247", parentdiffid="6384294"}

Reviewed By: @int3

Differential Revision: D1825313