Make socket/logs world-writable
Summary: This is the first step in letting HHVM talk to hh_server. In
order to have a single server per project, any user needs to be able to
talk to that server. Set umasks and permissions appropriately so that
can happen.

I'm pretty sure there aren't any security implications here. Anyone
could already write to `/tmp`. Swapping out the log file or socket
shouldn't get you anything, right -- the concern that comes to mind is a
random user who doesn't have permission to write to file X convincing a
running server to write to file X as its log file, but that shouldn't
happen, since we open the fd at start, after overwriting the contents.

{sync, type="parent", child="external", childrevid="35529", childdiffid="185139", parentdiffid="6870751"}

Reviewed By: @int3

Differential Revision: D1930684