Have clientBuild handle exception more robustly
Summary: It's possible that when the server is mismatched and old, when we sleep to wait
for it to restart, that the server takes a while to exit, meaning we can still
connect onto the old socket and then try to read from it. Of course the server
eventually exits, and we get a Sys_error "Connection reset by peer".

This changes the behavior of clientBuild to handle that case more gracefully
by waiting for the server to die by holding onto the currently connected socket
and reading from it until it is closed, then trying again.

Reviewed By: @int3

Differential Revision: D1825720