Better numa scheduling
Summary: When I wrote the code to handle binding threads
to numa nodes, I also updated the scheduling algorithm so
it would attempt to balance the traffic between the nodes.

ProxygenServer didn't enable this mode, but it looks like
LibeventServer tried but failed (the server was created
before we made the decision about numa scheduling, and so
the call to numa_num_nodes() always returned 1). It may
be that things were re-ordered since I wrote that.

This re-organizes things so that JobQueue takes care of it.
While doing that, I found that the code was swapping the
groups parameter (number of numa nodes) with the
queuedJobsReleaseRate parameter; which can't have been good.

Reviewed By: @bertmaher

Differential Revision: D1832496