Have hh_format's lexer eat up trailing spaces
Summary: This fixes an issue where we insert extra newlines.

For example, in handling a `case $a: \n` statement (note the extra
trailing space), we would insert a newline after `:`, and then the
remaining ` \n` would create an extra line below.

Seems simpler to fix this at the lexer rather than the parser level.

Reviewed By: @periodic1236

Differential Revision: D1829548