Use RegionDesc to determine merge points
Summary: The nextIsMerge function was treating all the blocks with the same
offset as the same, which is too pessimistic as it can claim some
blocks are merge points when they really aren't.  This diff rewrites
it so that it actually looks at the incoming region arcs into a block
to determine whether or not it's a merge point.

Reviewed By: @jdelong

Differential Revision: D1832287