Add a bit of a hack to guard relaxation to not retype LdStk<Gen> to StkElem
Summary: Normally this seems to work, but something has changed where we can
insert CheckFoos in front the of the EndGuards instruction (this might be the
multiple-copies-of-one-SrcKey in a region thing if it's the first block?).
These initial Checks are being optimized into unconditional jumps, and by the
type guard relaxation gets to a LdStk we might only know the stack slot was a
StkElem (code looking like this after guard relaxation: P19721829).  Since we
rely on gen-time typeParams to inform things of stack flavor information I
don't see a cleaner fix for now.

Reviewed By: @ottoni

Differential Revision: D1831997