Move a zend dom test to bad
Summary: It's passing "by accident", but apparently is misusing memory.  If you
add a few "$x = new stdclass" lines to it and turn on bcinterp:3 traces, it
starts failing.  Or (how I noticed it) if you change the JIT to punt RetC in
pseudomains it starts failing.  Presumably because either case is changing the
allocation pattern (it looks like it's a use after free of some sort, and the
DOMNode memory management is known to be broken...).

Reviewed By: @swtaarrs

Differential Revision: D1836266