Limit places where this::T can be used
Summary: The path dependent type analysis is in place to ensure soundness, but I
want to limit usage of 'this::T' style type constants to start. This
respects the current rules we have around usage of 'this'. I loosened
the rules to allow 'this' in one additional place, namely as function
parameters of static methods (including constructors).

Since I haven't moved 'this' to use the path dependent type machinery
yet, using 'this' in these positions is technically unsound. The options
were to thread through another flag, or move over 'this' to use path
dependent types in this diff. I elected not to go with either approach
since the risk of adding unsoundness in this way is rather minimal and I
plan on doing the refactor this week anyways.

Reviewed By: @jwatzman

Differential Revision: D1836571