Factor out get_defs
Summary: `get_defs` and `collect_defs` are almost identical, except that one does
`fold_left` and the other `fold_right`. `fold_right` gives nicer error messages
e.g. in the case where a function is declared twice in the same file, the error
will now say that the declaration with a larger line number is a duplicate.

`fold_right` is not tail-recursive, but in general the number of definitions
per file is relatively small.

I'm about to use this function for linting too, hence the factoring.

Reviewed By: @jwatzman

Differential Revision: D1836500