Fix Parser Bug for Type Constant
Summary: Didn't test my original test thoroughly enough, so the type constant
feature did not work if it appeared in a type parameter. This is because
it goes into scanner.cpp in this case to try to resolve '<' for a type
list. Simply needed to add a check in one place.

While adding test for this case I also discovered I wasn't constructing
the type annotation quite right for type constants either. Mainly
required shuffling some stuff in the parser and updating
TypeAnnotation::accessTypeName()

Reviewed By: @elgenie

Differential Revision: D1837502