Create an IntervalTimer extension
Summary: This extension allows you to set an interval timer that will
periodically run a PHP callback.  This feature makes it easy to write sampling
based profilers in PHP.

The implementation uses a timer thread to manage the time interval, as follows:

When `IntervalTimer::start()` is called, the extension:
- spawns a timer thread, and
- adds the IntervalTimer to a request local pool.

The timer thread does a `wait_for` on a condition variable for the requested
duration.  If it wakes due to timeout, it:
- marks the associated IntervalTimer as ready (`m_signaled`), and
- raise a surprise flag (`IntervalTimerFlag`), and
- goes back to sleep

When the request thread is surprised it checks the timer pool to see which
callbacks are ready, unmarks their ready state, and executes them.

When the request thread calls `stop()` (or when the IntervalTimer is destroyed,
or when the request ends), it notifies the condition variable to break out of
the sleep loop, and joins the timer thread.

Reviewed By: @billf

Differential Revision: D1829095