Move cycle detection from BlockableWaitHandle to WaitableWaitHandle
Summary: We are shifting towards a model where a single WaitHandle is allowed to
establish multiple dependencies via embedding multiple AsioBlockables.

In this world, BlockableWaitHandle becomes irrelevant. Let's move cycle
detection algorithm out of it. This algorithm assumes still assumes a single
dependency, but that's fine -- we will always assume only the first
non-finished dependency for the purposes of cycle detection and rerun the cycle
detection when the first non-finished dependency changes.

Reviewed By: @billf

Differential Revision: D1846214