Handle Tvar and Tunresolved for Taccess expansion
Summary: Currently I give up and emit and error if I encounter a Tvar or
Tunresolved while expanding the Taccess. Unfortunately this renders the
feature useless for all but the simplest examples because we create Tvar
and Tunresolved all the time for local type inference. To fix this I made two changes.

First: When initially expanding a Taccess I expand the root type and
distribute across any unresolved types. We then map over all the types
and create a new unresolved type as the result.

Second: I centralized the logic where we fill in the type hole present
for Expression Dependent Types into obj_get

Reviewed By: @jwatzman

Differential Revision: D1846802