zend initialization settings now work better
Summary: The const zend_ini_entry received from the PHP_INI_BEGIN..PHP_INI_END
data structure in the zend extension's source code is now copied
and associated with the system or user (per-request) CallbackMap.
This association is done with a new intermediate class UserIniData,
and mechanisms for registering a factory to produce (sub)classes of
UserIniData depending on the fundamental underlying semantics of the ini
setting.

The ini setting mechanisms in the zend compat layer have an instance
of a subclass of UserIniData to hold a pointer to the copy of the
zend_ini_entry, and that copy is freed when the request finishes or the
system finishes, perhaps with some "don't care" elision on a fast path
towards exit.

The string value assigned to a zend ini setting, as for example from
"ini_set", is now saved using estrndup, so that it can be regurgitated
directly when the zend ini setting is to be gotten with "ini_get". The
string is implicitly released at request finish time, or explictly
released on the
Closes https://github.com/facebook/hhvm/pull/4692

Reviewed By: @JoelMarcey

Differential Revision: D1797805