Change scoring function to match developers expectations better
Summary: This is one of the recurring complaints about hh_format - choosing:
  // 1
  $x =
    prep(
      TAO()->gen(
        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
      ),
    );
over
  // 2
  $x = prep(
    TAO()->gen(
      'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
    ),
  );
or
  // 3
  $x = prep(
    TAO()
      ->gen('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'),
  );
The way it happens is because rhs_assign in format_hack.ml branches into options 1 and 3 and the chooses one using keep_best scoring function. This function prefers solution with more "alignment", which is the number of consecutive lines starting with same character. I believe the reason for this is to prefer
  [..........]
  	->setX()
  	->setY()
  	->setZ()
over
  [..........]->setX()
  	->setY()
  	->setZ()
Because of this, it chooses solution 1, which has two consecutive lines starting with ')'. Looking at the examples from the real codebase it does more harm than good in most cases besides "->", so I change it to just look for this. Note: this will still sub-optimally format

  $x = AAA()
    ->setX('aaaaaaaaaaaaaaaaaaaaaaaaaa')
    ->setY('bbbbbbbbbbbbbbbbbbbbbbbbbb');

Because it chooses from
  $x =
    AAA()
      ->setX('aaaaaaaaaaaaaaaaaaaaaaaaaa')
      ->setY('bbbbbbbbbbbbbbbbbbbbbbbbbb')
and
  $x = AAA()->setX('aaaaaaaaaaaaaaaaaaaaaaaaaa')->setY(
    'bbbbbbbbbbbbbbbbbbbbbbbbbb',
  )

 - branching logic doesn't return solution that we would need at all, converging on some local maximum.

Reviewed By: @dlreeves, @elgenie

Differential Revision: D1842931