Loosen Interface Method Incompat Checks for Type Constants
Summary: Currently we do not enforce type constants as type constraints in the
runtime. Because of this it is useful if the implementer of an interface
can change the typehint declared in a parameter to the direct type
(without going through the type constant).

This is similar to the behavior we have for type parameters so I mostly
added the isTypeConstant() check where ever we had a isTypeVar() check.

One thing I looked into was having the runtime first implementers of an
interface to define some type hint if the interface's method was defined
using a type constant, but it seemed tricker to do in the runtime since
by the time we perform the check we have thrown away the distinction
between a parameter whose type is 'mixed' from a parameter with no type
hint.

I still like the idea of adding this enforcement so I will look into
adding this check to the runtime.

Reviewed By: @paroski

Differential Revision: D1849684