Fix a few incorrect cases in memory-effects
Summary: From audit while trying to debug a refcount problem.  (Neither of
these were the cause, and it seems unlikely they could cause problems for code
we currently generate, but they're definitely wrong.)  It was doing the same
thing for LdElem and Stelem which is obviously wrong.  And there was some
missing potential re-entry cases for class props.

Reviewed By: @edwinsmith

Differential Revision: D1850941