Simplifier tool
Summary: The simplifier tool is designed to reduce a php script to the simplest
case by comparing the behavior of a two hhvm binaries or configurations running
the given script. The script is then reduced until a minimal script is found
that still show different outputs from the two hhvms.

Reviewed By: @edwinsmith

Differential Revision: D1795166