Rewrite Enum classes in decl mode as well.
Summary: decl mode should be equivalent to having `// UNSAFE` in every function;
it should not affect the declarations exposed by the file.

Currently, in

  <?hh // decl

  class Foo extends Enum<string> {
    const BAR = 'bar';
  }

`Foo::BAR` is of type 'any', not string; but it is of type string if the
file is in partial mode. This commit makes it typed as a string in both
cases.

Reviewed By: @elgenie

Differential Revision: D1851070