Add hh_client --list-modes
Summary: This prints out a list of all the files and their associated
modes, in tab-separated columns.

This will be useful for e.g. converting decl files to partial mode, and
for keeping track of the number of files that we have for each
particular mode.

I think `--list-files` would be a better name, but alas, that is already taken.

Reviewed By: @elgenie

Differential Revision: D1849543