Refactor how hphpd computes autocomplete vectors
Summary: Right now it's passing around 6 vectors by reference and jamming Strings into
them.  Then it remembers it actually wanted std::string so it converts to that
and puts it in the real vector-array afterwards.  Make this a little saner by
just grabbing a reference to the vector-array in the first place.

Reviewed By: @bertmaher

Differential Revision: D1846142