Add type safety for stack offsets
Summary: We currently use raw ints to refer to stack offsets with various
frames of reference (e.g. relative to the FP, relative to SP, relative to
top-of-stack for the current bytecode, etc.). This patch replaces those
uses with single-member structs for each of the various frames of
reference, thus making the code clearer and giving us some type safety
(along with a couple bug fixes).

Reviewed By: @swtaarrs

Differential Revision: D1823186