Fix some bugs with reserved name restrictions
Summary: PHP5 and HHVM have a concept "reserved names" that cannot be used as class
names. For example, PHP5 reserves the names "self" and "parent" in all
namespaces to avoid confusion.

It turns out HHVM was not correctly matching PHP5's behavior for "self" and
"parent", and there were also some bugs with Hack-specific reserved names.
This diff fixes a couple of issues:

  1) HHVM now reserves "self" and "parent" in all namespaces to match PHP5
  behavior.
  2) HHVM now always reserves Hack-specific names in the HH namespace
  regardless of whether HH syntax is enabled.
  3) HHVM now always reserves Hack-specific names in the global namespace
  when HH syntax is enabled.

In addition to these fixes, this diff also rips out "TypeHintTypes" maps
from "compiler/analysis" which were redundant and bit-rotted, and instead
relies on nameToAnnotType() as the source of truth for reserved names. Also,
the error messages were improved when a class definition conflicts with
implicit auto-importation that happened earlier in the file / namespace
block.

Reviewed By: @elgenie

Differential Revision: D1873058