Fix ob_get_status() PHP compatibility
As documented and implemented in PHP PHP, type=0 is internal, and type=1
is for user handlers. This function got it the wrong way around.

Also reorder the array keys and add keys for chunk_size and buffer_used
per their undocumented PHP behaviour. When ob_start() is called without
arguments, give type=0 and name="default output handler", matching the
PHP behaviour. This gives a minimal diff when
zend/bad/tests/output/ob_get_status.php and zend/bad/tests/output/ob_013.php
are run. It now only complains about the omission of "flags" and
"buffer_size", both of which are implementation details that I don't
think should have been included in the PHP output.

Make ob_get_status_empty.php into a more specific test for #2690, so
that it does not restrict details of the return value of
ob_get_status() which are unrelated to that bug.