Make Lint module more independent of Errors
Summary: Re-using most of the errors logic was convenient to start with, but lint
has somewhat different requirements that seem to justify a separate
module. In particular, this allows us to add a field indicating lint severity,
and we may add an additional field to Lint.t for suggested autofixes. Both
these fields are supported by Arcanist.

Reviewed By: @elgenie

Differential Revision: D1856626