Add (and use) an AliasClass for MInstrState's TypedValues
Summary: Without this, the new reference count optimizer isn't able to assume
they have non-zero reference counts when they are loaded and decref'd, because
they don't have alias-analysis ids (so they must be treated conservatively).

Since the DecRefs of these tvRef guys are on tmps of type Gen, this means these
minstr decref sequences would reduce the lower bound of all tracked tmps by two
(one for each tvRef and tvRef2).  Moreover, I've noticed a few minstr sequences
don't even use the minstr state, but we never bothered to add them to needsMIS.
This diff is sufficient for load-elim to get rid of at least some of those
situations, too.

This diff is assuming that re-entering does not imply changing MInstrState in a
way visible to the current region (because it is saved and restored), so it
made sense to keep AMIStateAny out of AHeapAny, which involved some cleanup to
memory-effect's re-entry reporting.  It also assumes that no IR instructions
access MInstrState without being passed a misPtr().

Also: this diff fixes a bug in the memory effects for SetWithRefElem and
SetWithRefNewElem, which take the value to set as a pointer (anyone know why?),
so it should've been adding it in the may-load set but wasn't before the
new all_pointees thing.

Reviewed By: @swtaarrs

Differential Revision: D1863702