Split out ExtensionRegistry from Extension
Summary: The Extension class should be JUST the extension.
Leave the registry and callback dispatch to a separate thing.

This is mostly a refactoring diff, however I did make one
substantive change.

I moved HHVM_API_VERSION out of the Extension class
because I realized it wasn't actually working the
way it was intended to there.

Instead, I've added an ExtensionBuildInfo struct for
this and similar flags and updated the HHVM_GET_MODULE
macro to automatically populate a struct and return
the correct value.

I plan to followup this diff with one moving CompileSystemLib
to hphp/system/systemlib.{h,cpp} for obvious reasons. :)

{sync, type="parent", child="external", childrevid="33921", childdiffid="176367", parentdiffid="6576587"}

Reviewed By: @paulbiss

Differential Revision: D1866032