Clean up operand sizes in emitReffinessTest
Summary: Rather than emitting the smallest possible instructions for mask and
vals separately, decide on the operand size once and stick with it. This will
cause us to emit slightly larger code in some situations, but it's necessary
for the LLVM backend and any typechecking work we want to do in vasm.

Reviewed By: @markw65

Differential Revision: D1863940