Fix some issues with the perf counters
Summary: PERF_EVENT_IOC_RESET resets the current counter, but doesn't
reset the cumulative time that the counter has been active, nor the
cumulative time that its actually been running. So after a while,
the scaling factors we use when adjusting for multiplexing become
totally bogus.

I also noticed that if you have the inherit bit set (which I will
probably turn on in a later diff), the cumulative times for child
processes are not cleared either.

This just records the values at reset time so that everything should
match up.

Reviewed By: @bertmaher

Differential Revision: D1875881