Consume CURLOPT_SERVICE_NAME from HHVM for SPNEGO support
Summary: This diff depends on D1858769 which adds a new curl option CURLOPT_SERVICE_NAME to support SPNEGO from Curl. This is to consume the flag when HHVM is built within Facebook. The actual change to open sourced libcurl is being followed up separately.

This diff is against the HHVM release branch for testing purpose and will be moved off once the diff is accepted.

Reviewed By: @elgenie

Differential Revision: D1867057