Fold genZeroCheck=true into most of the cgDecRefWork thing
Summary: Unfortunately we can't delete cgDecRefStaticType until we delete
DecRefThis (#6317795), which limits how much we can clean up
cgCheckStaticBitAndDecRef.  (I don't want to port it to a new way because we're
going to delete it later.)  But this did clean up a fair bit of it.  I think
there is an even simpler function possible later...

Reviewed By: @ottoni

Differential Revision: D1879865