Kill some bogus asserts in FrameState, and fix predictedType tracking
Summary: When setting a type prediction, compute the predictedType as the
intersection of the given type and the known local/stack type.  This
allows us to properly enforce that the predictedType is a subtype of
the known local/stack type.

setBoxed{Local,Stk}Prediction were asserting that the type is Boxed,
but than may not hold in unreachable code or during fixed-point
computation. So kill those asserts.

Reviewed By: @jdelong

Differential Revision: D1870315