Relax default choice of asio context
Summary: Do not try to auto-import all dependencies into the current context. Instead,
select the highest possible context that satisfies dependencies when creating a
new WaitHandle object.

The dependencies are very likely to be already in the current context anyway
and if not, the first attempt to establish a dependency between an already
existing WaitHandle and the created WaitHandle will trigger enterContext().

Reviewed By: @jdelong

Differential Revision: D1870276