Collect post-conditions at IR-gen time
Summary: This diff moves the collection of post-conditions from being a
separate pass (based on FrameState) to IR-gen time.  This effectively
eliminates one of the FrameState uses in a post-IR-gen pass, thus
bringing us closer to have FrameState to be used only for IR-gen.
This diff also moves the code that actually collects the
post-conditions from IRUnit to FrameState, which seems more natural
since FrameState is the module actually providing the types of locals
& stack values.

Reviewed By: @jdelong

Differential Revision: D1880191