Some tweaks in assert detail
Summary: I wanted to be able to log the detail without aborting the program
(for debugging---e.g. print any unit that takes longer than X to compile), so
mostly this just makes it have a function that can do that.  Also make the
virtuals private and use some static member functions to expose stuff.  Also
move do_assert into HPHP::, and move bad_value from global scope to the use
site in minstr-effects.  The use site in the zend compat layer looks
potentially dangerous (it seems likely that it should be an always_assert
instead of a not_reached()?), but preserve the same behavior for now.

Reviewed By: @swtaarrs

Differential Revision: D1880666