Small fix to meth_caller invariant message
Summary: The error thrown by `meth_caller` when you try invoking the result on `null`
wasn't quite right.

`get_class` returns false when you pass it a non-object. And if you give it
null, it returns the name of the class you're in or raises an `E_WARNING`
level error should you be outside a class.

Reviewed By: @jwatzman

Differential Revision: D1880028