DOM: cleanup accessors to conform NPH signature
Summary: There was a useless conversion `ObjectData*` -> `Object` -> `ObjectData*` happening, because of old signature of DOM accessor methods. Thid diff brings direct signature of the NPH which should accept `ObjectData*`.

This will be used also in a fix if `CHECK_` methods (like `CHECK_DOC`) which were working with hardcoded `obj` var, and it wasn't easy to use this guard in other methods, that work directly with `this_` `ObjectData`.

{sync, type="parent", child="external", childrevid="34281", childdiffid="178593", parentdiffid="6652787"}

Reviewed By: @paulbiss

Differential Revision: D1882075