Add mechanisms to read glob-speced .ini and .hdf files.
Use these mechanisms to implement a policy of
reading all files that glob match:
  /etc/hhvm/php*.ini
  /etc/hhvm/config*.hdf
The knowledge of which files to load is found in 2 places.

This is an architectural change with ramifications I don't fully
understand.

This change will, however, make it easier to install and uninstall a .so
file and its corresponding .ini file.

Fix what seems to be a memory leak for error paths
out of the code implementing the php function "glob".