Set the memory limit default to 16GiB
Summary: It was suggested in #4725 that the memory limit was set to something a little more sensible than the current default of 9223372036854775807. This change sets the default memory limit to 1073741824 bytes (1GiB). Not sure if the test here is overkill, it is just testing a default after all...?

Potential risk - if users of HHVM rely on the extremely high default to "ignore" the memory limit, this could potentially cause breakages? I'm not sure what a sensible default might be (I guess, it is mostly subjective).
Closes https://github.com/facebook/hhvm/pull/4760

Reviewed By: @sgolemon

Differential Revision: D1821432