Treat 0.0s interval as run-once
Summary: It can be useful to have a non-interval timer as well (in JavaScript
this would be like setTimeout).  Setting the interval to 0 is a pretty concise
way of asking for this.

In particular I want to enable the pattern of re-setting a timer inside the
callback, to handle the case where the callback might be arbitrarily long
running, e.g., objprof.

Reviewed By: @alexmalyshev

Differential Revision: D1883965