Add a frame pointer to BCMarker
Summary: We need to know which frame we're on when explaining the memory
effects of instructions that can re-enter the VM.  Some of them take FramePtr
sources, and some don't.  This FramePtr doesn't count as a real source---it's
for use in the same way that the marker's spOffset is used.

Later on we'll hopefully restructure inlining not to involve redefining frames
(unless you're taking an exit inside the callee, which can probably be part of
the region-leaving instruction), and then we can remove this again.

Reviewed By: @swtaarrs

Differential Revision: D1897245