Clarify some code dealing with promoting intermediate props in hhbbc
Summary: When a prop in miProp doesn't have define, we shouldn't act like it
promotes to stdClass.  It currently doesn't cause anything to go wrong here,
because if it's not a defining minstr it'll fatal---the only affects of having
a Obj=stdClass instead of a Top in the PostProp base is whether things like
couldBeInThis later will say true, which we also only need in case we need to
promote to stdClass.  But it's clearer (and more consistent with the comment)
not to create the stdClass type if it isn't a defining prop.

Reviewed By: @ottoni

Differential Revision: D1888657