Add stack slots potentially smashed by re-entry to their GeneralEffects::kills
Summary: If an instruction can clobber the eval stack on re-entry, it's illegal
to load those slots without storing to them first, so we can put them in kills.
This lets us take AStackAny out of loads, too.  With this diff, only store-elim
can take advantage of this (but it's in the noise in perflab), but it'll be
useful for refcount opts in a bit.

Reviewed By: @swtaarrs

Differential Revision: D1897246