Fix reading Thrift's accelerated map when using compact protocol
Summary: The repro steps are currently somewhat complicated and are in the task but the problem is that we were returning an empty map object when deserializing maps.

The fix is something that I just copied from the binary protocol and it turned out to be working.
diffusion/FBCODE/browse/master/hphp/runtime/ext/thrift/binary.cpp;d412892d4ccf04a6e4de20e2b37be850129d003c$204-208

The sad part is that we do not have any unit tests. We have to do better.

@sgolemon - can you review this diff and explain why the former approach does not work? I see a few more instances of ArrayInit in this file and concerned that we have more bugs. Also if you have some tips how to unit test this change that would be awesome.

Reviewed By: @billf

Differential Revision: D1888241