Add ConstModifier to AuxUnion
Summary: We lose the distinction between a constant and type constant pretty
early on after parsing. This is the first step towards resolving that. I
modify the AuxUnion that we use to store m_isAbstractConst to instead
store a struct that contains the modifiers for a class constant.

This diff is meant to be a pure refactor so I didn't set the
ConstModifiers::m_isType field anywhere, but will do so in the follow up
diff.

Reviewed By: @elgenie

Differential Revision: D1890946