Fix a bug with abstract constants
Summary: hhbbc's analysis assumed that if two interfaces defined
the same constant, then instantiating the class would fatal at
runtime, so it ignored any class where this happened.

If one of the constants was abstract, however, the code was well
formed. But because hhbbc ignored the class, it also ignored any
parent child relationships, and, amongst other things, could set
AttrNoOverride on methods that were in fact overridden. This
then resulted in incorrect dispatch at runtime.

Reviewed By: @jdelong

Differential Revision: D1896817