Add copy propagation to vasm, let vxls allocate vmsp during traces
Summary: This completes the original goal of D1622952: vmsp is no longer forced
into rbx during traces. It's still synced from/to rbx at trace entry/exit, but
in the body of traces it lives in a normal SSA Vreg (it ends up in rbx most of
the time anyway due to copy hinting). I had to add a couple optimizations to
make this perf neutral, all of which live in vasm-copy.cpp:
- Convert "lea (%rx), %ry" to "copy %rx, %ry".
- Straightforward copy propagation on Vregs.
- Slightly less straightforward copy propagation between Vregs and PhysRegs.

The last point is where most of the complexity in this diff comes from. It
converts sequences like this:

  copy %rbx => %64
  ...
  copy %64 => %rbx
  fallbackcc ...
  ...
  copy %64 => %rbx
  fallbackcc ...

into this:

  copy %rbx => %64
  ...
  copy %rbx => %rbx
  fallbackcc ...
  ...
  copy %rbx => %rbx
  fallbackcc ...

The 2nd and 3rd copies were changed to use %rbx as a src rather than %64
because we were able to prove that %rbx still contained %64 at that point. This
eliminated all uses of %64, turning all of the copies into nop-copies or dead
code. In the cases where this isn't able to eliminate all uses of the Vreg, it
will still eliminate redundant copies from a Vreg into a PhysReg that already
contains the Vreg's value.

Reviewed By: @edwinsmith

Differential Revision: D1877614